Untangle types and typed ast
